fix(default-reporter): pass dylint, add trailing newline, harden reporter test
Builds on the restored ansi-diff differential renderer:

- pacquet: guarantee a trailing newline before diffing a frame, matching the
  TypeScript reporter. This keeps an interactive prompt on a fresh line below
  the frame instead of joined onto the last progress line, and leaves the
  differ's tracked cursor column at 0 so it stays in sync with the \r
  prepended on the next update (otherwise the inline diff computes relative
  moves from a stale column and corrupts in-place rendering).

- pacquet: move the differ's inline test module into src/diff/tests.rs with
  explicit imports and drop a stray trailing comma, so cargo dylint passes.

- TypeScript: locate the sticky lockfile verdict's first render with findIndex
  instead of assuming writes[0], and wrap each reporter test body in
  try/finally so the subscription is always stopped.
